+++ packages/atomic/src/pages/header.js
@@ -22,6 +22,7 @@ const links = [
   {href: '/examples/ipx.html', label: 'IPX'},
   {href: '/examples/genqa.html', label: 'Gen Q&A'},
   {href: '/examples/tabs.html', label: 'Tabs'},
+  {href: '/examples/commerce-website/homepage.html', label: 'Commerce Website'},
 ];
 
 const header = document.createElement('header');