+++ packages/quantic/cypress/e2e/default-2/generatedAnswer/generated-answer.cypress.ts
@@ -30,7 +30,7 @@ interface GeneratedAnswerOptions {
 
 let analyticsMode: 'legacy' | 'next' = 'legacy';
 const exampleTrackingId = 'tracking_id_123';
-const answerType = 'CRGA';
+const answerType = 'RGA';
 
 const GENERATED_ANSWER_DATA_KEY = 'coveo-generated-answer-data';
 const otherOption = 'other';
@@ -148,7 +148,8 @@ describe('quantic-generated-answer', () => {
           it('should perform a search query with the default rephrase button', () => {
             cy.wait(InterceptAliases.Search);
             Expect.searchQueryContainsCorrectRephraseOption(
-              defaultRephraseOption
+              defaultRephraseOption,
+              param.useCase === 'search' ? 'interfaceLoad' : 'searchboxSubmit'
             );
           });
 
@@ -159,7 +160,7 @@ describe('quantic-generated-answer', () => {
             );
           });
 
-          it('should display rephrase buttons', () => {
+          it('should display the rephrase buttons', () => {
             Expect.displayRephraseButtons(true);
             Expect.displayRephraseLabel(true);
           });
@@ -191,7 +192,8 @@ describe('quantic-generated-answer', () => {
               Expect.rephraseButtonIsSelected(conciseRephraseOption, false);
               Expect.rephraseButtonIsSelected(bulletRephraseOption, true);
               Expect.searchQueryContainsCorrectRephraseOption(
-                bulletRephraseOption
+                bulletRephraseOption,
+                param.useCase === 'search' ? 'interfaceLoad' : 'searchboxSubmit'
               );
             });
           });
@@ -278,7 +280,6 @@ describe('quantic-generated-answer', () => {
           describe(`when clicking the ${rephraseOption} rephrase button`, () => {
             const streamId = crypto.randomUUID();
             const secondStreamId = crypto.randomUUID();
-            const thirdStreamId = crypto.randomUUID();
 
             beforeEach(() => {
               mockSearchWithGeneratedAnswer(streamId, param.useCase);
@@ -289,7 +290,12 @@ describe('quantic-generated-answer', () => {
             it(`should send a new search query with the rephrase option ${option} as a parameter`, () => {
               scope('when loading the page', () => {
                 Expect.displayRephraseButtonWithLabel(rephraseOption);
-                Expect.rephraseButtonIsSelected(rephraseOption, false);
+                const expectedRephraseButtonSelected =
+                  option === defaultRephraseOption;
+                Expect.rephraseButtonIsSelected(
+                  rephraseOption,
+                  expectedRephraseButtonSelected
+                );
               });
 
               scope('when selecting the rephrase button', () => {
@@ -307,31 +313,14 @@ describe('quantic-generated-answer', () => {
                     Expect.displayRephraseButtonWithLabel(unselectedOption);
                     Expect.rephraseButtonIsSelected(unselectedOption, false);
                   });
-                Expect.searchQueryContainsCorrectRephraseOption(rephraseOption);
+                Expect.searchQueryContainsCorrectRephraseOption(
+                  rephraseOption,
+                  'rephraseGeneratedAnswer'
+                );
                 if (analyticsMode === 'legacy') {
                   Expect.logRephraseGeneratedAnswer(
                     rephraseOption,
                     secondStreamId
-                  );
-                }
-              });
-
-              scope('when unselecting the rephrase button', () => {
-                mockSearchWithGeneratedAnswer(thirdStreamId, param.useCase);
-                mockStreamResponse(thirdStreamId, genQaMessageTypePayload);
-
-                Actions.clickRephraseButton(rephraseOption);
-                rephraseOptions.forEach((unselectedOption) => {
-                  Expect.displayRephraseButtonWithLabel(unselectedOption);
-                  Expect.rephraseButtonIsSelected(unselectedOption, false);
-                });
-                Expect.searchQueryContainsCorrectRephraseOption(
-                  defaultRephraseOption
-                );
-                if (analyticsMode === 'legacy') {
-                  Expect.logRephraseGeneratedAnswer(
-                    defaultRephraseOption,
-                    thirdStreamId
                   );
                 }
               });
@@ -777,6 +766,7 @@ describe('quantic-generated-answer', () => {
                         },
                         citation: {
                           id: testCitations[hoveredCitationIndex].id,
+                          type: 'Source',
                         },
                       },
                       exampleTrackingId
@@ -810,6 +800,7 @@ describe('quantic-generated-answer', () => {
                       },
                       citation: {
                         id: testCitations[clickedCitationIndex].id,
+                        type: 'Source',
                       },
                     },
                     exampleTrackingId